+++ src/shaders/multiAnalyze1Calc.fs
@@ -1,13 +1,13 @@
-// #ifdef GL_FRAGMENT_PRECISION_HIGH
-// precision highp float;
-// #else
-// precision mediump float;
-// #endif
-
+#ifdef GL_FRAGMENT_PRECISION_HIGH
 precision highp float;
+#else
+precision mediump float;
+#endif
+
 precision highp sampler2D;
 
 #pragma glslify: rgbaToFloat = require(glsl-rgba-to-float)
+#pragma glslify: floatToRgba = require(glsl-float-to-rgba)
 #pragma glslify: isCloseEnough = require(./util/isCloseEnough.glsl)
 
 uniform sampler2D textureA;
@@ -28,9 +28,9 @@ void main() {
   bool aIsNodata = isCloseEnough(texelFloatA, nodataValue);
 
   if (aIsNodata || texelFloatA < filterLowA || texelFloatA > filterHighA) {
-    gl_FragColor = vec4(nodataValue);
+    gl_FragColor = floatToRgba(nodataValue, littleEndian);
   } else {
     float texelFloatFinal = texelFloatA * multiplierA;
-    gl_FragColor = vec4(texelFloatFinal);
+    gl_FragColor = floatToRgba(texelFloatFinal, littleEndian);
   }
 }